requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cmdkit==2.6.0
Django==3.2.5
django-elasticsearch-dsl==7.2.0
djangorestframework==3.12.4
elasticsearch==7.13.3
elasticsearch-dsl==7.3.0
importlib-metadata==4.6.1
Markdown==3.3.4
names==0.3.0
names-generator==0.1.0
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.6
zipp==3.5.0
